refactor: move more functions to forkchoice
