+++ lib/lambda_ethereum_consensus/p2p/gossip/attestation.ex
@@ -55,7 +55,7 @@ defmodule LambdaEthereumConsensus.P2P.Gossip.Attestation do
   end
 
   def publish_aggregate(%Types.SignedAggregateAndProof{} = signed_aggregate) do
-    fork_context = BeaconChain.get_fork_digest() |> Base.encode16(case: :lower)
+    fork_context = ForkChoice.get_fork_digest() |> Base.encode16(case: :lower)
     topic = "/eth2/#{fork_context}/beacon_aggregate_and_proof/ssz_snappy"
     {:ok, encoded} = SszEx.encode(signed_aggregate, Types.SignedAggregateAndProof)
     {:ok, message} = :snappyer.compress(encoded)
@@ -81,7 +81,7 @@ defmodule LambdaEthereumConsensus.P2P.Gossip.Attestation do
 
   defp topic(subnet_id) do
     # TODO: this doesn't take into account fork digest changes
-    fork_context = BeaconChain.get_fork_digest() |> Base.encode16(case: :lower)
+    fork_context = ForkChoice.get_fork_digest() |> Base.encode16(case: :lower)
     "/eth2/#{fork_context}/beacon_attestation_#{subnet_id}/ssz_snappy"
   end
