+++ lib/lambda_ethereum_consensus/p2p/gossip/operations_collector.ex
@@ -221,7 +221,7 @@ defmodule LambdaEthereumConsensus.P2P.Gossip.OperationsCollector do
   end
 
   defp topics() do
-    fork_context = BeaconChain.get_fork_digest() |> Base.encode16(case: :lower)
+    fork_context = ForkChoice.get_fork_digest() |> Base.encode16(case: :lower)
 
     topics =
       Enum.map(@topic_msgs, fn topic_msg ->
