+++ lib/lambda_ethereum_consensus/p2p/req_resp.ex
@@ -94,7 +94,7 @@ defmodule LambdaEthereumConsensus.P2P.ReqResp do
   @spec split_response(binary) :: {:ok, [binary()]} | {:error, String.t()} | {:error, Error.t()}
   def split_response(response_chunk) do
     # TODO: the fork_context should be computed depending on the block's slot
-    fork_context = BeaconChain.get_fork_digest()
+    fork_context = ForkChoice.get_fork_digest()
 
     case response_chunk do
       <<>> ->
