+++ lib/lambda_ethereum_consensus/validator/validator.ex
@@ -475,7 +475,7 @@ defmodule LambdaEthereumConsensus.Validator do
   defp publish_block(signed_block) do
     {:ok, ssz_encoded} = Ssz.to_ssz(signed_block)
     {:ok, encoded_msg} = :snappyer.compress(ssz_encoded)
-    fork_context = BeaconChain.get_fork_digest() |> Base.encode16(case: :lower)
+    fork_context = ForkChoice.get_fork_digest() |> Base.encode16(case: :lower)
 
     proposed_slot = signed_block.message.slot
 
@@ -494,7 +494,7 @@ defmodule LambdaEthereumConsensus.Validator do
   defp publish_sidecar(%Types.BlobSidecar{index: index} = sidecar) do
     {:ok, ssz_encoded} = Ssz.to_ssz(sidecar)
     {:ok, encoded_msg} = :snappyer.compress(ssz_encoded)
-    fork_context = BeaconChain.get_fork_digest() |> Base.encode16(case: :lower)
+    fork_context = ForkChoice.get_fork_digest() |> Base.encode16(case: :lower)
 
     subnet_id = compute_subnet_for_blob_sidecar(index)
